backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=1.10.0,<2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Core dependencies
pandas>=2.0.0
numpy>=1.24.0,<2.0.0
requests>=2.31.0
boto3>=1.28.0
botocore>=1.31.0

# Machine Learning
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.6.0,<3.7.0

# NLP and text processing
nltk>=3.8.0
yake>=0.4.8

# Data processing
json5>=0.9.0
python-dateutil>=2.8.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Fix compatibility issues
typing-extensions>=4.0.0

# Enhanced RAG with transformer models
accelerate>=0.20.0
tokenizers>=0.13.0

# Fine-tuning dependencies
peft>=0.4.0
datasets>=2.14.0
bitsandbytes>=0.41.0

# Visualization
plotly==5.17.0
python-json-logger==2.0.7
